+++ source/deprecated/omni.isaac.physics_utilities/omni/isaac/physics_utilities/__init__.py
@@ -1,0 +1,20 @@
+# Copyright (c) 2018-2023, NVIDIA CORPORATION. All rights reserved.
+#
+# NVIDIA CORPORATION and its licensors retain all intellectual property
+# and proprietary rights in and to this software, related documentation
+# and any modifications thereto. Any use, reproduction, disclosure or
+# distribution of this software and related documentation without an express
+# license agreement from NVIDIA CORPORATION is strictly prohibited.
+#
+
+import carb
+
+old_extension_name = "omni.isaac.physics_utilities"
+new_extension_name = "isaacsim.util.physics"
+
+# Provide deprecation warning to user
+carb.log_warn(
+    f"{old_extension_name} has been deprecated in favor of {new_extension_name}. Please update your code accordingly."
+)
+
+from isaacsim.util.physics import *